+++ src/main/java/net/sodiumstudio/befriendmobs/entity/ai/goal/BefriendedGoal.java
@@ -2,6 +2,7 @@ package net.sodiumstudio.befriendmobs.entity.ai.goal;
 
 import java.util.HashSet;
 import java.util.Random;
+import java.util.function.Predicate;
 
 import javax.annotation.Nullable;
 
@@ -21,8 +22,8 @@ public abstract class BefriendedGoal extends Goal implements IBefriendedGoal {
 	protected static final BefriendedAIState WANDER = BefriendedAIState.WANDER;
 	
 	protected IBefriendedMob mob = null;
-	private HashSet<BefriendedAIState> allowedStates = new HashSet<BefriendedAIState>();
-	private boolean isBlocked = false;
+	protected HashSet<BefriendedAIState> allowedStates = new HashSet<BefriendedAIState>();
+	protected boolean isBlocked = false;
 	protected Random rnd = new Random();
 	/**
 	 *  If it's more than 0, when checking if this goal should be executed (only on start, not on check continue using),
@@ -33,6 +34,11 @@ public abstract class BefriendedGoal extends Goal implements IBefriendedGoal {
 	 * If true, this goal will require the mob's owner is in the same level to run.
 	 */
 	protected boolean requireOwnerPresent = true;
+	/**
+	 * Additional condition to start this goal. (Not checked on checking continue to use)
+	 */
+	protected Predicate<BefriendedGoal> startCondition = null;
+	
 	
 	@Deprecated
 	public BefriendedGoal()
@@ -137,6 +143,8 @@ public abstract class BefriendedGoal extends Goal implements IBefriendedGoal {
 			throw new RuntimeException("Illegal method call: checkCanUse() method cannot call canUse() method inside, otherwise an infinite loop will occur. To get super class' check, call checkCanUse().");
 		if (mob == null || requireOwnerPresent && !mob.isOwnerPresent())
 			return false;
+		if (startCondition != null && !startCondition.test(this))
+			return false;
 		if (skipChance > 0 && rnd.nextDouble() < skipChance)
 			return false;
 		BefriendedGoalCheckCanUseEvent event = new BefriendedGoalCheckCanUseEvent(this, BefriendedGoalCheckCanUseEvent.Phase.CAN_USE);
@@ -169,7 +177,7 @@ public abstract class BefriendedGoal extends Goal implements IBefriendedGoal {
 			return false;
 		return checkCanContinueToUse();
 	}
-
+	
 	/**
 	 * Get skip chance of this goal.
 	 * <p> If it has skip chance > 0, it will have a chance to be directly skipped when checking if to start, allowing goals below to be executed.
@@ -189,6 +197,26 @@ public abstract class BefriendedGoal extends Goal implements IBefriendedGoal {
 	public <T extends BefriendedGoal> T setSkipChance(double value)
 	{
 		this.skipChance = value;
+		return (T)this;
+	}
+	
+	/**
+	 * Set additional start condition of this goal. This method returns the goal itself, and use template class to specify the subclass to cast.
+	 */
+	@SuppressWarnings("unchecked")
+	public <T extends BefriendedGoal> T setStartCondition(Predicate<BefriendedGoal> condition)
+	{
+		this.startCondition = condition;
+		return (T)this;
+	}
+	
+	/**
+	 * Set if requires the owner to be present to start this goal.
+	 */
+	@SuppressWarnings("unchecked")
+	public <T extends BefriendedGoal> T setRequireOwnerPresent(boolean value)
+	{
+		this.requireOwnerPresent = value;
 		return (T)this;
 	}