+++ src/main/java/net/sodiumstudio/befriendmobs/entity/ai/goal/BefriendedMoveGoal.java
@@ -3,7 +3,7 @@ package net.sodiumstudio.befriendmobs.entity.ai.goal;
 import java.util.function.Predicate;
 
 import net.minecraft.world.level.block.Blocks;
-import net.sodiumstudio.befriendmobs.entity.ai.IBefriendedAmphibious;
+import net.sodiumstudio.befriendmobs.entity.befriended.IBefriendedAmphibious;
 import net.sodiumstudio.befriendmobs.entity.befriended.IBefriendedMob;
 
 /**