+++ src/main/java/net/sodiumstudio/befriendmobs/entity/ai/goal/BefriendedTargetGoal.java
@@ -1,11 +1,12 @@
 package net.sodiumstudio.befriendmobs.entity.ai.goal;
 
 import java.util.HashSet;
+import java.util.Random;
+import java.util.function.Predicate;
 
 import javax.annotation.Nullable;
 
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.target.TargetGoal;
 import net.minecraftforge.common.MinecraftForge;
@@ -24,10 +25,20 @@ public abstract class BefriendedTargetGoal extends TargetGoal implements IBefrie
 	protected IBefriendedMob mob = null;
 	protected HashSet<BefriendedAIState> allowedStates = new HashSet<BefriendedAIState>();
 	protected boolean isBlocked = false;
+	protected Random rnd = new Random();
+	/**
+	 *  If it's more than 0, when checking if this goal should be executed (only on start, not on check continue using),
+	 *  it will has chance to be directly skipped, allowing goals below to be executed.
+	 */
+	protected double skipChance = 0;
 	/**
 	 * If true, this goal will require the mob's owner is in the same level to run.
 	 */
 	protected boolean requireOwnerPresent = true;
+	/**
+	 * Additional condition to start this goal. (Not checked on checking continue to use)
+	 */
+	protected Predicate<BefriendedTargetGoal> startCondition = null;
 	
 	public BefriendedTargetGoal(IBefriendedMob mob, boolean mustSee)
 	{
@@ -133,6 +144,10 @@ public abstract class BefriendedTargetGoal extends TargetGoal implements IBefrie
 			throw new RuntimeException("Illegal method call: checkCanUse() method cannot call canUse() method inside, otherwise an infinite loop will occur. To get super class' check, call checkCanUse().");
 		if (mob == null || requireOwnerPresent && !mob.isOwnerPresent())
 			return false;
+		if (startCondition != null && !startCondition.test(this))
+			return false;
+		if (skipChance > 0 && rnd.nextDouble() < skipChance)
+			return false;
 		BefriendedGoalCheckCanUseEvent event = new BefriendedGoalCheckCanUseEvent(this, BefriendedGoalCheckCanUseEvent.Phase.CAN_USE);
 		MinecraftForge.EVENT_BUS.post(event);
 		if (event.getManualSetValue().isPresent())
@@ -169,4 +184,47 @@ public abstract class BefriendedTargetGoal extends TargetGoal implements IBefrie
 	{
 		return true;
 	}
+	
+	/**
+	 * Get skip chance of this goal.
+	 * <p> If it has skip chance > 0, it will have a chance to be directly skipped when checking if to start, allowing goals below to be executed.
+	 * When it's skipped, it won't post {@link BefriendedGoalCheckCanUseEvent}.
+	 */
+	public double getSkipChance()
+	{
+		return skipChance;
+	}
+	
+	/**
+	 * Set skip chance of this goal. This method returns the goal itself, and use template class to specify the subclass to cast.
+	 * <p> If it has skip chance > 0, it will have a chance to be directly skipped when checking if to start, allowing goals below to be executed.
+	 * When it's skipped, it won't post {@link BefriendedGoalCheckCanUseEvent}.
+	 */
+	@SuppressWarnings("unchecked")
+	public <T extends BefriendedTargetGoal> T setSkipChance(double value)
+	{
+		this.skipChance = value;
+		return (T)this;
+	}
+	
+	/**
+	 * Set additional start condition of this goal. This method returns the goal itself, and use template class to specify the subclass to cast.
+	 */
+	@SuppressWarnings("unchecked")
+	public <T extends BefriendedTargetGoal> T setStartCondition(Predicate<BefriendedTargetGoal> condition)
+	{
+		this.startCondition = condition;
+		return (T)this;
+	}
+	
+	/**
+	 * Set if requires the owner to be present to start this goal.
+	 */
+	@SuppressWarnings("unchecked")
+	public <T extends BefriendedTargetGoal> T setRequireOwnerPresent(boolean value)
+	{
+		this.requireOwnerPresent = value;
+		return (T)this;
+	}
+	
 }