+++ src/main/java/net/sodiumstudio/befriendmobs/entity/ai/goal/BefriendedTargetGoal.java
@@ -10,7 +10,7 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.target.TargetGoal;
 import net.minecraftforge.common.MinecraftForge;
-import net.sodiumstudio.befriendmobs.bmevents.entity.BefriendedGoalCheckCanUseEvent;
+import net.sodiumstudio.befriendmobs.bmevents.entity.ai.BefriendedGoalCheckCanUseEvent;
 import net.sodiumstudio.befriendmobs.entity.ai.BefriendedAIState;
 import net.sodiumstudio.befriendmobs.entity.befriended.IBefriendedMob;